task-common/src/main/java/io/pivotal/scdf/demo/task/common/model/Publisher.java
package io.pivotal.scdf.demo.task.common.model;

public class Publisher {

    private String pubId;
    private String pubName;
    private String address;
    private String city;
    private String state;

    @Override
    public String toString() {
        return "Publisher{" +
                "pubId='" + pubId + '\'' +
                ", pubName='" + pubName + '\'' +
                ", address='" + address + '\'' +
                ", city='" + city + '\'' +
                ", state='" + state + '\'' +
                '}';
    }

    public String getPubId() {
        return pubId;
    }

    public void setPubId(String pubId) {
        this.pubId = pubId;
    }

    public String getPubName() {
        return pubName;
    }

    public void setPubName(String pubName) {
        this.pubName = pubName;
    }

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public String getCity() {
        return city;
    }

    public void setCity(String city) {
        this.city = city;
    }

    public String getState() {
        return state;
    }

    public void setState(String state) {
        this.state = state;
    }
}
